fixed postinstall script

+++ build_macos_pkg.sh
@@ -3,6 +3,14 @@
 PLATFORM_DIR="$PWD/platforms/macos"
 PKGBUILD_DIR="$PLATFORM_DIR/rvc.dist"
 
+readonly brew="/usr/local/bin/brew"
+if [ ! -x "${brew}" ]; then
+	echo "Couldn't execute '${brew}'. (Hint: install Homebrew 'https://brew.sh/')"
+	exit 1
+fi
+
+readonly OPENVPN_DIR="$(${brew} --prefix openvpn)"
+
 # remove old dist directory
 rm -r "$PKGBUILD_DIR"
 
@@ -18,10 +26,13 @@ OPENSSL_DIR=$(brew --prefix openssl)
 make
 make install
 
+# copy openvpn binrary to dist directory
+install -c $OPENVPN_DIR/sbin/openvpn $PKGBUILD_DIR/sbin
+
 # build package
 pkgbuild \
 	--root "$PKGBUILD_DIR" \
 	--identifier com.ribose.rvc \
-	--install-location /opt/rvc.dist \
+	--install-location /tmp/rvc.dist \
 	--scripts "$PLATFORM_DIR/scripts" \
 	"$PLATFORM_DIR/rvc.pkg"
